fix long to int conversion
Removed unnecessary long to int conversion in LoanApplicationTerms.java
Fixed narrow calculation in LoanSchedularServiceImpl.java by adding explicit cast to double
Removed unnecessary long to int conversions in LoanWritePlatformServiceJpaRepositoryImpl.java
Removed unnecessary long to int conversion in AdHocScheduledJobRunnerServiceImpl.java
Added explicit cast to double for long value in AbstractWorkbookPopulator.java

+++ fineract-provider/src/main/java/org/apache/fineract/portfolio/loanaccount/service/LoanSchedularServiceImpl.java
@@ -232,7 +232,7 @@ public class LoanSchedularServiceImpl implements LoanSchedularService {
         // get the size of current paginated dataset
         int size = loanIds.size();
         // calculate the batch size
-        double toGetCeilValue = size / threadPoolSize;
+        double toGetCeilValue = size / (double) threadPoolSize;
         batchSize = (int) Math.ceil(toGetCeilValue);
 
         if (batchSize == 0) {
